stocks_api/notifications/requirements.txt
flask-restful
flask
requests
Flask-Mail
flask-sqlalchemy
mysqlclient
sqlalchemy_utils
marshmallow
marshmallow_sqlalchemy